+++ src/main/kotlin/blackjack/model/strategy/CardShuffler.kt
@@ -1,6 +1,6 @@
 package blackjack.model.strategy
 
-import blackjack.model.domain.Card
+import blackjack.model.domain.card.Card
 
 interface CardShuffler {
     fun spread(cards: List<Card>): List<Card>
