+++ src/main/kotlin/blackjack/model/strategy/TrueShuffle.kt
@@ -1,6 +1,6 @@
 package blackjack.model.strategy
 
-import blackjack.model.domain.Card
+import blackjack.model.domain.card.Card
 
 class TrueShuffle : CardShuffler {
     override fun spread(cards: List<Card>): List<Card> {
