+++ src/test/kotlin/blackjack/model/domain/DealerTest.kt
@@ -1,5 +1,9 @@
 package blackjack.model.domain
 
+import blackjack.model.domain.card.Card
+import blackjack.model.domain.card.CardNumber
+import blackjack.model.domain.card.Shape
+import blackjack.model.domain.participant.Dealer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
