+++ src/test/kotlin/blackjack/model/domain/ParticipantStatusTest.kt
@@ -1,5 +1,6 @@
 package blackjack.model.domain
 
+import blackjack.model.domain.participant.ParticipantStatus
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
